+++ lib/Dialect/CKKS/Transforms/Passes.h
@@ -2,6 +2,8 @@
 #define LIB_DIALECT_CKKS_TRANSFORMS_PASSES_H_
 
 // IWYU pragma: begin_keep
+#include "lib/Dialect/CKKS/IR/CKKSDialect.h"
+#include "lib/Dialect/CKKS/Transforms/CKKSToLWE.h"
 #include "lib/Dialect/CKKS/Transforms/DecomposeKeySwitch.h"
 #include "lib/Dialect/CKKS/Transforms/DecomposeRelinearize.h"
 #include "mlir/include/mlir/Pass/Pass.h"          // from @llvm-project
