+++ lib/Dialect/CKKS/Transforms/Passes.td
@@ -19,4 +19,21 @@ def DecomposeKeySwitch : Pass<"ckks-decompose-keyswitch"> {
   let dependentDialects = ["mlir::heir::ckks::CKKSDialect", "tensor::TensorDialect"];
 }
 
+def CKKSToLWE : Pass<"ckks-to-lwe"> {
+  let summary = "Lower `ckks` to `lwe` dialect.";
+
+  let description = [{
+    This pass lowers the `ckks` dialect to `lwe` dialect.
+    Note that some scheme specific ops (e.g., rescale) that
+    have no direct analogue in the `lwe` dialect are left unchanged.
+    TODO (#1193): support both "common" and "full" lwe lowering
+  }];
+
+  let dependentDialects = [
+    "mlir::heir::ckks::CKKSDialect",
+    "mlir::heir::lwe::LWEDialect",
+    "mlir::tensor::TensorDialect",
+  ];
+}
+
 #endif  // LIB_DIALECT_CKKS_TRANSFORMS_PASSES_TD_
