+++ lib/Pipelines/BUILD
@@ -98,7 +98,7 @@ cc_library(
     deps = [
         ":PipelineRegistration",
         "@heir//lib/Dialect/BGV/Conversions/BGVToLWE",
-        "@heir//lib/Dialect/CKKS/Conversions/CKKSToLWE",
+        "@heir//lib/Dialect/CKKS/Transforms:CKKSToLWE",
         "@heir//lib/Dialect/Debug/Transforms",
         "@heir//lib/Dialect/Debug/Transforms:ValidateNames",
         "@heir//lib/Dialect/LWE/Conversions/LWEToLattigo",
